Update sma_crossover.cpp
Use const reference for passing vectors to functions:
When passing a vector to a function, using a const reference can be more efficient than passing by value. This way, the function can access the vector's elements without copying them.

Use range-based loops:
Instead of using a for loop with an index variable, you can use a range-based for loop to iterate over the elements of a vector.

Avoid unnecessary copies:
Instead of creating new vectors inside the calculateSMA function, we can pass iterators to the start and end of the subvector that we want to compute the moving average for.